+++ demo/app/boot/app.config.ts
@@ -1,5 +1,4 @@
 import {ApplicationConfig, importProvidersFrom} from '@angular/core';
-import {provideAnimations} from '@angular/platform-browser/animations';
 import {HotkeyModule} from 'angular2-hotkeys';
 
 import {appProviders} from './app.providers';
@@ -7,12 +6,11 @@ import {appProviders} from './app.providers';
 /**
  * Application configuration for browser
  */
-export const appConfig: ApplicationConfig = 
+export const appConfig: ApplicationConfig =
 {
     providers:
     [
         ...appProviders,
-        provideAnimations(),
         importProvidersFrom(HotkeyModule.forRoot(
         {
             cheatSheetCloseEsc: true