+++ src/modules/logger/components/index.ts
@@ -1,2 +1,1 @@
 export * from './console/console.component';
-export * from './console/console.component.animations';